xen/grant-table: add gnttab_try_end_foreign_access()
commit 72103f08cde854334ec48d30f33d58a642d68344 from
https://github.com/altera-opensource/linux-socfpga.git

[upstream commit 6b1775f26a2da2b05a6dc8ec2b5d14e9a4701a1a]

Add a new grant table function gnttab_try_end_foreign_access(), which
will remove and free a grant if it is not in use.

Its main use case is to either free a grant if it is no longer in use,
or to take some other action if it is still in use. This other action
can be an error exit, or (e.g. in the case of blkfront persistent grant
feature) some special handling.

This is CVE-2022-23036, CVE-2022-23038 / part of XSA-396.

Reported-by: Demi Marie Obenour <demi@invisiblethingslab.com>
Signed-off-by: Juergen Gross <jgross@suse.com>
Reviewed-by: Jan Beulich <jbeulich@suse.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>