net: phy: 10G mode for fixed PHYs
commit d6dc53b8f1190c8bbc15e918ce38ae867babe3fa from
git@git.assembla.com:cavium/WindRiver.linux.git

The port 9 & 10 on a marvell DSA switch needs to be connected in
fixed-link mode to get 10G speed enabled. This mode uses an emumated
PHY that exposes standard registers. However, multiGigabit PHYs tend
to use standardised C45 registers to exposes these modes, which aren't
supported in swphy.

We could add these modes, and that will probably be done in mainline,
but in a way that is very difficult to backport. It implies reworking
the way we represent link-modes, which impacts all drivers in
net/ethernet and net/phy.

For now, add a way to simulate this feature by hardcoding the
phy_interface_modes that would need such PHYs to be emulated.

Signed-off-by: Sujeet Baranwal <sbaranwal@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <SunilKovvuri.Goutham@cavium.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Change-Id: Ic9f3f793e59a08120754c4ec890f14cc759f7305
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/74387
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>