octeontx2-bphy-netdev: re-enable cpri gpint when interface is stopped
commit 84900d17af85f01906f54e0bc6675ecbcd31fdb6 from
git@git.assembla.com:cavium/WindRiver.linux.git

There is a problem in netdev related to the enabling/disabling of
the RX GP interrupt when the interface where napi polling was made
is stopped while GP interrupt was disabled from the last poll.

Polling from all the interfaces belonging to the same CPRI DMA block
is always done from the 1st enabled interface in the CPRI DMA block.
When the interface is stopped, it's napi is disabled and ideally when
next packets are are received polling will be done from the next enabled
interface. The problem happens when napi is stopped before it actually
re-enables the GP interrupts and in this case reception for the whole
CPRI block is stopped.

The patch fixes the issue by re-enabling the RX GPINT when the interface
is disabled.

Change-Id: I98dc5d86749142766eb7f6ec65e883d15e074dd2
Signed-off-by: Baha Mesleh <baha.mesleh@nokia.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/77399
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
(cherry picked from commit 95085811cfdffabe03298ac9df660686431d8958)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/77439
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>