octeontx2-pf: revert the ptp phc implementation to use timecounter
commit 879c521c9fdd099fb01229498b2e2f0b5d159eb9 from
git@git.assembla.com:cavium/WindRiver.linux.git

To support 1-step PTP, the PHC implementation was changed to use
the low-level PTP hardware block registers directly instead of
of timecounter approach because timecounter maintains it's own
offset. However, when adjtime is called, the driver performs a
register read-modify-write operation, introducing an offset of
300+ns when testing synchronization with an external timestamper.
To avoid this issue, the implementation reverts back to the
timecounter approach.

Change-Id: Iaedd658506691c0bc8d5b7ab20c136cef428c0ea
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/77146
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>