cn10k: Add workaround for MRVL TAD Errata-38891
commit 2f571de1829479757b1b4b2175ef9fc85fdd3c29 from
git@git.assembla.com:cavium/WindRiver.linux.git

An issue exists whereby if one core (core A) is accessing a cache
line and a certain set of race cases occur when a second core
(Core B) is executing a DC CVAC, DC CVAP or DC CVADP to the same
cache line, the second core B may respond to the Last Level Cache
(TAD's) request for data with an unexpected transaction type. This
will cause the TAD to ignore the response, response will never come
to Core A, so as soon as Core A reaches an instruction that requires
that response to complete it will hang. Core B and other cores will
remain responsive.

Workaround is to replace below cache maintenance instructions:
1. DC CVAC with DC CIVAC.
2. DC CVAP with DC CIVAC; DC CVAP.
3. DC CVADP with DC CIVAC; DC CVAP (DC CVAP is equivalent to
                                    CVADP for our system).

Any userspace execution of these instruction are configured to
trap to kernel and replace as above. Kernel usages of these
instruction are replaced using alternatives and compile time
kernel config CONFIG_ARM64_WORKAROUND_MARVELL_38891 option

Change-Id: I7ac88de80b83b030c8eaae00d984a546655afd09
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/78551
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
[1. Drop some codes which are always exist.
 2. By commit 4ef333b2d10680b5d966a733ed7171f72164fcd5, change:
    force_signal_inject(SIGILL, ILL_ILLOPC, regs->pc) to
    force_signal_inject(SIGILL, ILL_ILLOPC, regs->pc, 0);
]
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>