octeontx2-af: Fix pf bitmap for RVU VFs
commit 12a1521f6bb30ff0215c2a14217c9dad295e56cb from
git@git.assembla.com:cavium/WindRiver.linux.git

pf_bitmap is set based on the "RVU_PRIV_PFX_CFG[ena]" bit.
But this is applicable only for the PFs. But current code
use same for RVU VFs due which some valid VFs are not
getting initilized. This patch fixes it.

fixes: 7694279d66379 ("octeontx2-af: Skip PFs if not enabled")
Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: I6ff43608a37a859d30b6cc5be680982638cc258e
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/81326
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>