tick/nohz_full: Ensure quiet_vmstat() is called on exit to user-mode when the idle tick is stopped
commit fb275315168bfddaf528e4dc3492d49178688405 from
git@git.assembla.com:cavium/WindRiver.linux.git

This patch ensures CPU-specific vmstat differentials do not remain
when the scheduling-tick is stopped and before exiting to user-mode
in the context of nohz_full only.

A trivial test program was used to determine the impact of the proposed
changes and under vanilla. The mlock(2) and munlock(2) system calls was
used solely to modify vmstat item 'NR_MLOCK'. The following is an average
count of CPU-cycles across the aforementioned system calls:

                                  Vanilla                 Modified

  Cycles per syscall              8461                    8690    (+2.6%)

Signed-off-by: Aaron Tomlin <atomlin@redhat.com>
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
[Fixed backporting issues]

Change-Id: Ia2e179c802acd6a20a6301e221677993803f5015
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/87321
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>