octeontx2-bphy-netdev: cnf10k: check psm queue space for ptp queue
commit d311929d973fc0000817bc6f23ede5ac887f7de0 from
git@git.assembla.com:cavium/WindRiver.linux.git

Because PTP packets are enqueued to HW one after the other by the driver,
the existing PTP 2-step handling code does not check the PSM queue space
before submitting a PTP job because the PSM queue will be always free to
enqueue. When using PTP 1-step mode, the PTP packets are timestamped on
the fly in HW. The driver enqueues PTP packets without waiting for the
PTP completion interrupt. So the driver has been changed to check the
PSM queue space for the PTP psm queue. This has no effect on the
2-step mode. Also modified netif_wake_queue to check both the PSM
queues.

Change-Id: Ib90a0fa7956f38a24bd0aa38eff6b47c9634a5d9
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89657
Reviewed-by: Rakesh Babu Saladi <rsaladi2@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>