drivers: i2c: octeon: Avoid potential infinite loop
commit 795b9347609314c5d346a4b2eba5e1100a6010ca from
git@git.assembla.com:cavium/WindRiver.linux.git

The time_left variable is used and set to 1000, which is not
being checked in while loop condition - this could trigger as
infinite loop in case the test flag is not set. Adding time_left
check in while condition as a way to exit the loop in that case.

Signed-off-by: Piyush Malgujar <pmalgujar@marvell.com>
Change-Id: Ib0b1fbbe5ae83b63b88372f77b15ea33b26c38a0
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/89919
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>